Fix classname comparison in update_nodeping_check
This is a workaround for a bug in the upstream nodeping.py library:
title() returns the checktype name with the first letter in uppercase
but the checktypes use in fact PascalCase.

For example, 'Http' works fine but 'HttpAdv' breaks. The workaround is
to convert both to uppercase to do an case insensitive comparison.